requirements-test.txt
# Testing Dependencies
pytest==8.3.3
pytest-cov==6.0.0
pytest-flask==1.3.0
pytest-mock==3.14.0
coverage==7.6.7
faker==33.0.0
